Add detekt custom rule and lint custom rule

+++ data/network/src/main/kotlin/com/grappim/network/model/waybill/WaybillProductDTO.kt
@@ -50,7 +50,7 @@ data class PartialWaybillProductDTO(
     @Serializable(BigDecimalSerializer::class)
     val sellingPrice: BigDecimal,
     @SerialName("waybillId")
-    val waybillId: Int,
+    val waybillId: Long,
     @SerialName("productId")
     val productId: Long
 )
