+++ domain/src/main/kotlin/com/grappim/domain/interactor/waybill/UpdateWaybillProductUseCase.kt
@@ -15,7 +15,7 @@ class UpdateWaybillProductUseCase @Inject constructor(
 ) : FlowUseCase<UpdateWaybillProductUseCase.Params, BigDecimal>(ioDispatcher) {
 
     data class Params(
-        val waybillId: Int,
+        val waybillId: Long,
         val barcode: String,
         val name: String,
         val purchasePrice: BigDecimal,
